% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudo_bulk}
\alias{pseudo_bulk}
\title{pseudo_bulk}
\format{
A SummarizedExperiment object.
}
\usage{
data(pseudo_bulk)
}
\description{
This object was saved only because leading to a strange failure of the github action,
while working perfectly in the local environment
}
\details{
A SummarizedExperiment object
}
\keyword{datasets}
